# services/app-management/api/app_management/openapi.yaml
openapi: 3.0.3
info:
  title: 6DAiNAS-OS App Management API
  version: v2
  description: |
    <picture>
        <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/6dainas/6dainas-os/main/logo/6dainas_banner_dark_800px.png">
        <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/6dainas/6dainas-os/main/logo/6dainas_banner_light_800px.png">
        <img alt="6DAiNAS-OS" src="https://raw.githubusercontent.com/6dainas/6dainas-os/main/logo/6dainas_banner_light_800px.png">
    </picture>

    6DAiNAS-OS App Management manages app lifecycle, including installing, uninstalling, starting, stopping, setting, etc.

    For issues and discussions, please visit the [GitHub repository](https://github.com/6dainas/6dainas-os) or join [our Discord](https://discord.gg/knqAbbBbeX).
servers:
  - url: /v2/app_management
...
security:
  - access_token: []
paths:
...
  /container/{id}:
    patch:
      summary: Recreate the container app
      description: |
        TODO - add description
      operationId: recreateContainerByID
      tags:
        - Container methods
      parameters:
        - $ref: "#/components/parameters/ContainerID"
        - $ref: "#/components/parameters/PullLatestImage"
        - $ref: "#/components/parameters/Force"
      responses:
        "200":
          $ref: "#/components/responses/ContainerRecreateOK"
        "404":
          $ref: "#/components/responses/ResponseNotFound"
        "500":
          $ref: "#/components/responses/ResponseInternalServerError"
...
components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization
  parameters:
...
    ContainerID:
      name: id
      description: Container ID in the Docker runtime
      in: path
      required: true
      schema:
        type: string
      example: 09e70f4a7dd3fc43de6a1bc40d45c79f857544e13679e5b1a47409034a7ce621
...
    PullLatestImage:
      name: pull
      description: Try to pull the latest image from the registry
      in: query
      schema:
        type: boolean
        default: false
    Force:
      name: force
      in: query
      schema:
        type: boolean
        default: false
...
  responses:
...
    ResponseInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Internal Server Error"
    ResponseNotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Not Found"
...
    ContainerRecreateOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BaseResponse"
...
  schemas:
    BaseResponse:
      properties:
        message:
          description: message returned by server side if there is any
          type: string
          example: ""
...